% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiagnosisHeatmap.R
\name{DiagnosisHeatmap}
\alias{DiagnosisHeatmap}
\title{Visualize the Diagnosis Heatmap}
\usage{
DiagnosisHeatmap()
}
\arguments{
\item{No}{empty}
}
\description{
Visualize the Diagnosis Heatmap
}
\examples{
DiagnosisHeatmap()
}
